(T): add test for prelude completion
