+++ src/test/resources/org/rust/lang/core/completion/fixtures/stdlib/prelude.rs
@@ -1,0 +1,3 @@
+fn main() {
+    dr<caret>
+}
