+++ src/test/resources/org/rust/lang/core/completion/fixtures/stdlib/prelude_after.rs
@@ -1,0 +1,3 @@
+fn main() {
+    drop
+}
